s3_upload/requirements.txt
mysql-connector-python
py-dateutil